Added functionality for functions and multiple files.
Calling, returning, and defining functions added as well as a goto fuction. Also added support for multiple files to be translated at once. Requires the directory path as an argument instead of the file path. Currently I don't think this works on a single .vm file unless it is the only file in the directory.